% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matrix_which}
\alias{matrix_which}
\title{Get coordinate information for matrices}
\usage{
matrix_which(mat)
}
\arguments{
\item{mat}{Numeric matrix}
}
\value{
Data frame containing each 2D coordinate and corresponding value
}
\description{
This function compresses matrix outputs into a data frame containing coordinates for each value.
This is used for compressing the binary and cell mask outputs in the generate_GP_Masks function.
}
