% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{separable_GP}
\alias{separable_GP}
\title{Generating predictive mean from image matrix}
\usage{
separable_GP(output_mat, parameters)
}
\arguments{
\item{output_mat}{Image matrix}

\item{parameters}{Range and nugget parameters estimated using \code{separable_GP_param_est()}}
}
\value{
Returns a list containing:
\item{predmean_mat}{Predictive mean matrix}
\item{grad1}{Horizontal gradient}
\item{grad2}{Vertical gradient}
\item{grad_magnitude}{Gradient magnitude}
\item{param}{Gaussian parameters}
}
\description{
This function generates the predictive mean for an image matrix using a Gaussian process.
}
