% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{separable_GP_param_est}
\alias{separable_GP_param_est}
\title{Estimate nugget and range parameters for predictive mean of an image matrix}
\usage{
separable_GP_param_est(output_mat)
}
\arguments{
\item{output_mat}{Image matrix}
}
\value{
Returns a list containing the range and nugget parameters
}
\description{
This function estimates the nugget and range parameters of the Gaussian process for
constructing the predictive mean of the segmented paper via Eigendecomposition.
}
