% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_GP_Masks}
\alias{generate_GP_Masks}
\title{Generate cell masks by fast Gaussian processes}
\usage{
generate_GP_Masks(
  file_path,
  delta = 0.01,
  nugget = T,
  middle_threshold = 0.15,
  boundary_threshold = 0.05
)
}
\arguments{
\item{file_path}{File path for cell image to segment}

\item{delta}{Step size for percentages to be tested for criterion 1 (default is 0.01)}

\item{middle_threshold}{Size threshold for filtering object masks not touching the image boundary (removes anything smaller than threshold * mean mask size); default is 0.15}

\item{boundary_threshold}{Size threshold for filtering object masks touching the image boundary (removes anything smaller than threshold * mean mask size); default is 0.05}
}
\value{
Returns a list containing:
\item{ori_images}{Original version of sectioned image}
\item{processed_images}{Predictive mean for each image section}
\item{crit_1_opt_thresholds}{Optimal threshold by criterion 1 for each image section}
\item{connected_parts_count}{Number of unique objects for each image section after thresholding before watershed}
\item{outliers}{IDs for thresholded image section that contains unusually large object counts}
\item{combined_predmean}{Predictive mean matrix for entire image}
\item{combined_thresholded1}{Binary matrix for entire image before watershed}
\item{GP_masks}{Cell masks for entire image after thresholding binary matrix, with each cell mask having a unique ID}
}
\description{
This function generates object masks for cell microscopy images using fast Gaussian processes for smoothing,
a data-driven threshold for foreground/background segmentation, and watershed for separating touching cell objects.
Note that this function divides the original image into sections for more robust processing.
}
